requirements.txt
Django==1.11.3
djangorestframework==3.6.3

psycopg2==2.7.3
Fabric==1.13.1
Fabric3
opbeat==3.5.2
redis==2.10.5
celery==4.2.1
zeep==2.5.0

mutagen==1.38
django-silk==1.0.0

kavenegar==1.1.0
django-admin-rangefilter==0.3.6
google-auth==1.5.0
jdatetime==2.1.0
django-constance==2.2.0
mido==1.2.8
ddtrace==0.12.1
requests==2.18.4
requests-toolbelt==0.8.0
django-user-agents==0.3.2
ua-parser==0.8.0
django-user-agents==0.3.2
user-agents==1.1.0
